1.change test case 2.use fprint_escape instead of write_escape_string
